Tablet reading
Summary:
- Load tablet footer.
- Load tablet indexes.
- Add resources for testing.
- Find in tablets (puts it all together).

Test Plan: unit tests

Reviewers: peter, tberman

Reviewed By: peter

CC: jimmy, justin

Differential Revision: http://phabricator.thefactory.com/D716

+++ csharp/TheFactory.Datastore/src/Block.cs
@@ -27,7 +27,7 @@ namespace TheFactory.Datastore {
             }
             var end = start + length - 4;
             stream.Seek(end, SeekOrigin.Begin);
-            numRestarts = stream.ReadInt();
+            numRestarts = (int)stream.ReadInt();
             return (int)numRestarts;
         }
 
@@ -126,7 +126,7 @@ namespace TheFactory.Datastore {
 
             private long ReadRestartOffset(int index) {
                 stream.Seek(restarts + (index * 4), SeekOrigin.Begin);
-                return stream.ReadInt();
+                return (long)stream.ReadInt();
             }
 
             private byte[] ReadKey(long offset) {
